Code maintenance for Qt 5: - Added new options to retroshare-gui.pro for Qt 5 - Added new file QtVersion.h with macros to compile with Qt 4 and Qt 5   - QHeaderView::setResizeMode -> QHeaderView::setSectionResizeMode   - QHeaderView::setMovable -> QHeaderView::setSectionsMovable
git-svn-id: http://svn.code.sf.net/p/retroshare/code/trunk@6849 b45a01b8-16f6-495d-af2f-9b41ad6348cc

+++ retroshare-gui/src/gui/AboutDialog.cpp
@@ -221,7 +221,7 @@ void AWidget::timerEvent(QTimerEvent* e) {
     page ^= 1;
 
     if (qrand()  % 128 == 0) {
-        int r = 3 + qRound(qrand() * 4 / RAND_MAX);
+        int r = 3 + qRound((double) qrand() * 4 / RAND_MAX);
         int h = 300 + qrand() * 200 / RAND_MAX;
         int x = 1 + r + qrand()%(image1.width() -2*r-1);
         int y = 1 + r + qrand()%(image1.height()-2*r-1);
